@@ -4,7 +4,7 @@ version: 1.1.0
 homepage: https://github.com/yako-dev/flutter-full-screen-menu
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
   flutter:
